+++ src/main/java/net/sf/mpxj/primavera/p3/P3DatabaseReader.java
@@ -387,6 +387,19 @@ public final class P3DatabaseReader implements ProjectReader
             task.setWBS(((Task) parent).getWBS());
          }
 
+         int percentComplete = task.getPercentageComplete().intValue();
+         if (percentComplete < 0)
+         {
+            task.setPercentageComplete(Integer.valueOf(0));
+         }
+         else
+         {
+            if (percentComplete > 100)
+            {
+               task.setPercentageComplete(Integer.valueOf(100));
+            }
+         }
+         
          int flag = row.getInteger("ACTUAL_START_OR_CONSTRAINT_FLAG").intValue();
          if (flag != 0)
          {