+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 8.0.3 (git master)
 * Improve handling of zero value durations, costs and units from certain MPP files.
+* Improve percent complete calculation for certain XER file and P6 Database schedules.
+* Improve percent complete calculation for certain P3 schedules.
 
 ## 8.0.2 (16/02/2020)
 * Improve handling of zero duration tasks read from Phoenix files.